[AUTO PR] [AUTO PR] Merge rc 3.68 to master

+++ features/travel/hotel/src/main/java/com/tokopedia/hotel/booking/data/model/HotelPropertyData.kt
@@ -8,7 +8,7 @@ class HotelPropertyData {
 
     @SerializedName("propertyID")
     @Expose
-    val propertyID: Int = 0
+    val propertyID: Long = 0
 
     @SerializedName("name")
     @Expose
